Refactoring KB
